page/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "page"
version = "0.0.1"
description = "Webpage for romancse.com"
authors = [{ name = "Roman", email = "iamromandev@gmail.com" }]
readme = "README.md"
requires-python = "==3.13.2"

dependencies = [
    "loguru>=0.7.3",
    "django>=5.2",
    "django-environ>=0.12.0",
    "django-cors-headers>=4.7.0",
    "django-browser-reload>=1.18.0",
    "dj-static>=0.0.6",
    "django-soft-delete>=1.0.18",
    "django-taggit>=6.1.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "django-bootstrap-icons>=0.9.0",
    "django-avatar>=8.0.1",
    "pymysql>=1.1.1",
    "django-phonenumber-field[phonenumbers]>=8.0.0",
    "prometheus-client>=0.21.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.11.2",
    "mypy>=1.15.0",
]

[tool.uv]
package = false
